+++ BlazorWasm/BinaryTool/BinaryTool/Dom/DomKind.cs
@@ -1,0 +1,11 @@
+﻿namespace BinaryTool.Dom;
+
+public enum DomKind
+{
+    Unknown,
+    Object,
+    Array,
+    Value,
+
+    Error = -1,
+}
