+++ BlazorWasm/BinaryTool/BinaryTool/Dom/DomSpan.cs
@@ -1,0 +1,3 @@
+﻿namespace BinaryTool.Dom;
+
+public record DomSpan(DomKind Kind, Range Range, object? Value, object? Key, int Length);
