+++ BlazorWasm/BinaryTool/BinaryTool/Dom/GlobalUsings.cs
@@ -1,0 +1,2 @@
+global using List = System.Collections.Generic.List<object?>;
+global using Map = System.Collections.Generic.Dictionary<object, object?>;
