+++ BlazorWasm/BinaryTool/BinaryTool/Dom/ILoader.cs
@@ -1,0 +1,7 @@
+namespace BinaryTool.Dom;
+
+public interface ILoader
+{
+    string Description { get; }
+    List<DomSpan> Parse(byte[] data);
+}
